+++ pha-admin/build.gradle
@@ -31,6 +31,7 @@ dependencies {
 	implementation 'org.thymeleaf.extras:thymeleaf-extras-springsecurity5'
 	implementation 'org.springdoc:springdoc-openapi-ui:1.5.2'
 	implementation "org.apache.commons:commons-lang3:3.9"
+	implementation 'net.kaczmarzyk:specification-arg-resolver:2.6.2'
 	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	runtimeOnly 'mysql:mysql-connector-java'
