--- pha-admin/src/main/java/com/store/pharmacy/Company/repository/CompanyRepository.java
@@ -1,9 +1,0 @@
-package com.store.pharmacy.Company.repository;
-
-import com.store.pharmacy.Company.model.Company;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface CompanyRepository extends JpaRepository<Company, Long> {
-}
