+++ pha-admin/src/main/java/com/store/pharmacy/company/repository/CompanyRepository.java
@@ -1,0 +1,10 @@
+package com.store.pharmacy.company.repository;
+
+import com.store.pharmacy.company.model.Company;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface CompanyRepository extends JpaRepository<Company, Long>, JpaSpecificationExecutor<Company> {
+}
