+++ pha-admin/src/main/java/com/store/pharmacy/utils/Paging.java
@@ -1,0 +1,21 @@
+package com.store.pharmacy.utils;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+@AllArgsConstructor
+public class Paging {
+
+    private Long count;
+
+    private Long pageNumber;
+
+    private Long pageSize;
+
+    private Long pageOffset;
+
+    private Long pageTotal;
+}
