+++ pha-admin/src/main/java/com/store/pharmacy/utils/PharmacyConstant.java
@@ -1,7 +1,10 @@
 package com.store.pharmacy.utils;
 
 public class PharmacyConstant {
-	
-	public static final String ROLE_USER = "ROLE_USER";
 
+    public static final String ROLE_USER = "ROLE_USER";
+
+    public static final String PAGE_NUMBER = "Page-Number";
+
+    public static final String PAGE_SIZE = "Page-Size";
 }
