+++ pha-admin/src/main/resources/application.properties
@@ -20,9 +20,9 @@ spring.jpa.properties.hibernate.dialect= org.hibernate.dialect.MySQL5InnoDBDiale
 
 # Database
 spring.datasource.driver-class-name= com.mysql.cj.jdbc.Driver
-spring.datasource.url= jdbc:mysql://us-cdbr-east-04.cleardb.com:3306/heroku_408321431540760?zeroDateTimeBehavior=convertToNull&useUnicode=yes&characterEncoding=UTF-8&serverTimezone=${app.timeZone}
-spring.datasource.username= b941c455c16229
-spring.datasource.password= b23c9af1
+spring.datasource.url= jdbc:mysql://localhost:3306/phastore?zeroDateTimeBehavior=convertToNull&useUnicode=yes&characterEncoding=UTF-8&serverTimezone=${app.timeZone}
+spring.datasource.username= root
+spring.datasource.password= root
 
 # Mail server
 spring.mail.host=smtp.gmail.com
